% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_links.R
\name{extract_links}
\alias{extract_links}
\title{Extract Links from HTML}
\usage{
extract_links(x, NA.omit = FALSE)
}
\arguments{
\item{x}{Character. A vector of filenames or URLs.}

\item{NA.omit}{Logical. Whether to return results with or without NAs included. Defaults to FALSE.}
}
\value{
A vector of hyperlinks (for a single HTML document) or a named list of vectors (for a vector of document names).
}
\description{
Extracts hyperlinks from one or more HTML documents by reading a HTML markup, selecting all <a> nodes and returning their 'href' attributes as a vector. Robust implementation that returns 'NA' on error.
}
