% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffreqtable.R
\name{ffreqtable}
\alias{ffreqtable}
\title{Fast Frequency Tables
Creates frequency tables for an arbitrary number of variables. It can return them as a list, write them to an arbitrary folder on disk as CSV files (with an optional prefix or return kable tables that are designed to work well with render() and LaTeX. It is based on data.table and is therefore capable of quickly processing massive data sets.}
\usage{
ffreqtable(
  x,
  varlist = names(x),
  sumrow = TRUE,
  out.list = TRUE,
  out.kable = FALSE,
  out.csv = FALSE,
  out.dir = "./",
  prefix = "",
  align = "r"
)
}
\arguments{
\item{x}{A data.frame or data.table.}

\item{varlist}{\link{Optional} Character. A vector of variable names to construct tables for. Defaults to all variables.}

\item{sumrow}{\link{Optional} Logical. Whether to add a summary row. Defaults to TRUE.}

\item{out.list}{\link{Optional} Logical. Whether to output the frequency tables as a list. Defaults to TRUE. Returns NULL otherwise.}

\item{out.kable}{\link{Optional} Logical. Whether to return kable tables. Defaults to FALSE.}

\item{out.csv}{\link{Optional} Logical. Whether to write CSV files (one per variable) to disk. Defaults to FALSE.}

\item{out.dir}{\link{Optional} Character. The target directory for writing CSV files. Defaults to the current R working directory.}

\item{prefix}{\link{Optional} A string to be prefixed to the filename of each CSV file. Default is not to add a string and just to output the variable name as the name of the CSV file.}

\item{align}{\link{Optional} Alignment of table columns passed to kable. Default is "r". Note that your options passed must work for a five-column layout.}
}
\value{
Returns a list of frequency tables for each variable by default. Different out.* arguments can add CSV and kable output.
}
\description{
To show the kable output in render() you must add the following chunk option: "results = 'asis'"
}
