% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihashes.R
\name{multihashes}
\alias{multihashes}
\title{Computation of SHA2 and SHA3 Hashes
Computes SHA2-256 and SHA3-512 hashes for a single file. It returns a data frame with the file name, SHA2-256 hashe and SHA3-512 hash.  The function requires the existence of the openssl (RPM) library on the system.}
\usage{
multihashes(x)
}
\arguments{
\item{x}{Path to file.
\subsection{Required: OpenSSL System Library}{

The function requires the existence of the openssl (RPM) library on the system.
}}
}
\description{
Computation of SHA2 and SHA3 Hashes
Computes SHA2-256 and SHA3-512 hashes for a single file. It returns a data frame with the file name, SHA2-256 hashe and SHA3-512 hash.  The function requires the existence of the openssl (RPM) library on the system.
}
