% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_txt.R
\name{pdf_to_txt}
\alias{pdf_to_txt}
\title{\subsection{Conversion of PDF to TXT}{

Extracts text from PDF files and writes the result to disk as TXT files. Resulting TXT files have the same filename as the original document (only the extension is modified).
}}
\usage{
pdf_to_txt(x)
}
\arguments{
\item{x}{Path to a PDF file.}
}
\value{
A TXT file on disk with the same basename as the original PDF file.
}
\description{
\subsection{Conversion of PDF to TXT}{

Extracts text from PDF files and writes the result to disk as TXT files. Resulting TXT files have the same filename as the original document (only the extension is modified).
}
}
