% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_empty_NA.R
\name{vec_empty_NA}
\alias{vec_empty_NA}
\title{Replaces empty elements in vector with NA}
\usage{
vec_empty_NA(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Replaces empty elements in vector with NA
}
