% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_iso.R
\name{year_iso}
\alias{year_iso}
\title{\subsection{year.iso: Transform Two-Digit Years to Four-Digit Years}{

This function transforms two-digit years (YY) to four-digit years compliant with ISO-8601 (YYYY). It is based on the assumption that two-digit years above a certain boundary year belong to the reference century and years at or below that boundary belong to the following century.
}}
\usage{
year_iso(x, boundary = 50, century = 1900)
}
\arguments{
\item{x}{A vector of two-digit years.}

\item{boundary}{The boundary year. Defaults to 50 (= 1950).}

\item{century}{The reference century. Defaults to the 1900s (20th century).}
}
\value{
A vector of four-digit years.
}
\description{
\subsection{year.iso: Transform Two-Digit Years to Four-Digit Years}{

This function transforms two-digit years (YY) to four-digit years compliant with ISO-8601 (YYYY). It is based on the assumption that two-digit years above a certain boundary year belong to the reference century and years at or below that boundary belong to the following century.
}
}
\examples{
year_iso(91)
year_iso(c(13, 45, 36, 57, 68))
}
