% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_pdf_to_txt.R
\name{future_pdf_to_txt}
\alias{future_pdf_to_txt}
\title{\subsection{Parallelized Conversion of PDF to TXT}{

Extracts text from PDF files and writes the result to disk as TXT files. Parallel implementation with the future package. Resulting TXT files have the same filename as the original document (only the extension is modified).
}}
\usage{
future_pdf_to_txt(x)
}
\arguments{
\item{x}{A vector of PDF filenames.}
}
\value{
A set of TXT files on disk with the same basename as the original PDF files. Invisible return in R session.
}
\description{
Please note that you must declare your own future evaluation strategy. By default the function will be evaluated sequentially. Use plan(multisession, workers = n) on Windows or plan(multicore, workers = n) on Linux/Mac, where n stands for the number of CPU cores you wish to use. Due to the need to read/write to the disk the function may not work properly on high-performance clusters.
}
